+++ fe/src/contexts/HeaderProvider.tsx
@@ -6,10 +6,11 @@ interface IHeaderState {
   isLogin: boolean;
   isDarkMode: boolean;
   userInfo: MemberType | null;
+  accessToken: string | null;
 }
 
 type Action =
-  | { type: 'LOGIN'; userInfo: MemberType }
+  | { type: 'LOGIN'; userInfo: MemberType; accessToken: string }
   | { type: 'LOGOUT' }
   | { type: 'THEME_TOGGLE' }
   | { type: 'REFRESH_TOKEN'; accessToken: string };
@@ -20,6 +21,7 @@ const initHeaderState: IHeaderState = {
   isLogin: false,
   isDarkMode: Boolean(localStorage.getItem('isDarkMode')) || false,
   userInfo: null,
+  accessToken: null,
 };
 
 /*
@@ -34,6 +36,7 @@ const initHeaderStateForDefaultPage: IHeaderState = {
     name: '익조',
     profileUrl: 'https://avatars.githubusercontent.com/u/82401504?v=4',
   },
+  accessToken: 'fakeToken',
 };
 //
 
@@ -47,12 +50,14 @@ function reducer(state: IHeaderState, action: Action): IHeaderState {
         ...state,
         isLogin: true,
         userInfo: action.userInfo,
+        accessToken: action.accessToken,
       };
     case 'LOGOUT':
       return {
         ...state,
         isLogin: false,
         userInfo: null,
+        accessToken: null,
       };
     case 'THEME_TOGGLE': {
       const toggleData = !state.isDarkMode;
@@ -62,6 +67,11 @@ function reducer(state: IHeaderState, action: Action): IHeaderState {
         isDarkMode: toggleData,
       };
     }
+    case 'REFRESH_TOKEN':
+      return {
+        ...state,
+        accessToken: action.accessToken,
+      };
     default:
       throw new Error('Unhandled action');
   }
